# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Flash Sale Order & Inventory API
  version: 0.1.0
  description: |
    Design-first MVP contract for products, inventory, and concurrency-safe
    ordering. Authentication and payment are intentionally outside this phase.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security: []
tags:
  - name: Products
    description: Product catalogue and management.
  - name: Inventory
    description: Inventory visibility and reasoned adjustments.
  - name: Orders
    description: Concurrency-safe ordering and order queries.
paths:
  /api/v1/products:
    get:
      tags: [Products]
      summary: List products
      operationId: listProducts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/ProductSort"
      responses:
        "200":
          description: A page of products.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [Products]
      summary: Create a product and its initial inventory
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created.
          headers:
            Location:
              description: URI of the created product.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/products/{productId}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [Products]
      summary: Get a product
      operationId: getProduct
      responses:
        "200":
          description: Product details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      tags: [Products]
      summary: Partially update a product
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/inventories:
    get:
      tags: [Inventory]
      summary: List inventory
      operationId: listInventory
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/InventorySort"
      responses:
        "200":
          description: A page of inventory records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryPage"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/inventories/{productId}/adjustments:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    post:
      tags: [Inventory]
      summary: Adjust available inventory
      description: Applies a non-zero signed delta and records its business reason.
      operationId: adjustInventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustInventoryRequest"
      responses:
        "200":
          description: Updated inventory.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inventory"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "409":
          $ref: "#/components/responses/ConflictProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Place an order
      description: |
        Creates an order and decrements inventory in one transaction. Repeating
        the same key and payload for a customer returns the existing order.
      operationId: placeOrder
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Existing Order returned for an identical idempotent replay.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "201":
          description: Order accepted.
          headers:
            Location:
              description: URI of the accepted order.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "409":
          $ref: "#/components/responses/ConflictProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        default:
          $ref: "#/components/responses/Problem"
    get:
      tags: [Orders]
      summary: List accepted orders
      operationId: listOrders
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/OrderSort"
        - $ref: "#/components/parameters/OrderProductId"
        - $ref: "#/components/parameters/OrderCustomerId"
        - $ref: "#/components/parameters/CreatedFrom"
        - $ref: "#/components/parameters/CreatedTo"
      responses:
        "200":
          description: A page of orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "500":
          $ref: "#/components/responses/InternalProblem"
        default:
          $ref: "#/components/responses/Problem"
components:
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      description: Product identity.
      schema:
        type: integer
        format: int64
        minimum: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key, unique within one customer.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    Page:
      name: page
      in: query
      description: Zero-based page number.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Records per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ProductSort:
      name: sort
      in: query
      description: Allowlisted sort expression; ID is always the stable tie-breaker.
      schema:
        type: string
        enum:
          - "id,asc"
          - "id,desc"
          - "name,asc"
          - "name,desc"
          - "price,asc"
          - "price,desc"
          - "createdAt,asc"
          - "createdAt,desc"
        default: "id,asc"
    InventorySort:
      name: sort
      in: query
      description: Allowlisted sort expression; product ID is the stable tie-breaker.
      schema:
        type: string
        enum:
          - "productId,asc"
          - "productId,desc"
          - "productName,asc"
          - "productName,desc"
          - "availableQuantity,asc"
          - "availableQuantity,desc"
        default: "productId,asc"
    OrderSort:
      name: sort
      in: query
      description: Allowlisted sort expression; ID is always the stable tie-breaker.
      schema:
        type: string
        enum: ["id,asc", "id,desc", "createdAt,asc", "createdAt,desc"]
        default: "createdAt,desc"
    OrderProductId:
      name: productId
      in: query
      description: Return only orders for this Product.
      schema:
        type: integer
        format: int64
        minimum: 1
    OrderCustomerId:
      name: customerId
      in: query
      description: Return only orders placed by this Customer.
      schema:
        type: string
        format: uuid
    CreatedFrom:
      name: createdFrom
      in: query
      description: Inclusive lower bound for Order creation time.
      schema:
        type: string
        format: date-time
    CreatedTo:
      name: createdTo
      in: query
      description: |
        Exclusive upper bound for Order creation time. When both bounds are
        supplied, this value must be later than `createdFrom`.
      schema:
        type: string
        format: date-time
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationProblem:
      description: Request validation failed with code `VALIDATION_FAILED`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblem"
    NotFoundProblem:
      description: Product was not found (`PRODUCT_NOT_FOUND`).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ConflictProblem:
      description: |
        Business conflict such as `PRODUCT_NOT_AVAILABLE`,
        `INSUFFICIENT_STOCK`, or `IDEMPOTENCY_KEY_REUSED`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalProblem:
      description: Unexpected failure with code `INTERNAL_ERROR`; internal details are omitted.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/InternalProblem"
  schemas:
    Product:
      type: object
      required: [id, name, price, currency, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          description: Leading and trailing whitespace is removed before validation.
          minLength: 1
          maxLength: 120
          pattern: '^\S(?:.*\S)?$'
        description:
          type: [string, "null"]
          maxLength: 1000
        price:
          type: number
          multipleOf: 0.01
          exclusiveMinimum: 0
        currency:
          type: string
          const: VND
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      additionalProperties: false
      required: [name, price, active, initialInventory]
      properties:
        name:
          type: string
          description: Must not contain leading or trailing whitespace.
          minLength: 1
          maxLength: 120
          pattern: '^\S(?:.*\S)?$'
        description:
          type: [string, "null"]
          maxLength: 1000
        price:
          type: number
          multipleOf: 0.01
          exclusiveMinimum: 0
        active:
          type: boolean
        initialInventory:
          type: integer
          minimum: 0
          maximum: 1000000
    UpdateProductRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          description: Must not contain leading or trailing whitespace.
          minLength: 1
          maxLength: 120
          pattern: '^\S(?:.*\S)?$'
        description:
          type: [string, "null"]
          maxLength: 1000
        price:
          type: number
          multipleOf: 0.01
          exclusiveMinimum: 0
        active:
          type: boolean
    Inventory:
      type: object
      required: [productId, productName, availableQuantity, updatedAt]
      properties:
        productId:
          type: integer
          format: int64
        productName:
          type: string
        availableQuantity:
          type: integer
          format: int64
          minimum: 0
        updatedAt:
          type: string
          format: date-time
    AdjustInventoryRequest:
      type: object
      additionalProperties: false
      required: [quantityDelta, reason]
      properties:
        quantityDelta:
          type: integer
          format: int64
          not:
            const: 0
        reason:
          type: string
          minLength: 3
          maxLength: 200
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [customerId, productId, quantity]
      properties:
        customerId:
          type: string
          format: uuid
        productId:
          type: integer
          format: int64
          minimum: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 5
    Order:
      type: object
      required:
        [id, customerId, productId, productName, quantity, unitPrice, currency, totalAmount, createdAt]
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: string
          format: uuid
        productId:
          type: integer
          format: int64
        productName:
          type: string
        quantity:
          type: integer
          minimum: 1
          maximum: 5
        unitPrice:
          type: number
          multipleOf: 0.01
          exclusiveMinimum: 0
        currency:
          type: string
          const: VND
        totalAmount:
          type: number
          multipleOf: 0.01
          exclusiveMinimum: 0
        createdAt:
          type: string
          format: date-time
    PageMetadata:
      type: object
      required: [number, size, totalElements, totalPages]
      properties:
        number:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ProductPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        page:
          $ref: "#/components/schemas/PageMetadata"
    InventoryPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Inventory"
        page:
          $ref: "#/components/schemas/PageMetadata"
    OrderPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        page:
          $ref: "#/components/schemas/PageMetadata"
    Problem:
      type: object
      description: RFC 9457 Problem Details with stable application extensions.
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        code:
          type: string
          description: Stable machine-readable application error code.
        traceId:
          type: string
          description: Correlation identifier for diagnostics.
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          required: [fieldErrors]
          properties:
            fieldErrors:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"
    InternalProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            code:
              type: string
              const: INTERNAL_ERROR
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string